+++ src/app/button/button.demo.ts
@@ -1,8 +1,10 @@
 import {
   Component,
   OnInit,
-  ChangeDetectionStrategy
+  ChangeDetectionStrategy,
+  ViewChild
 } from '@angular/core';
+import { SohoButtonComponent } from 'ids-enterprise-ng';
 
 @Component({
   selector: 'app-button-demo',
@@ -12,9 +14,31 @@ import {
 export class ButtonDemoComponent implements OnInit {
   public shouldSayHi = false;
 
+  @ViewChild('primaryGenerative', { static: true }) primaryGenerative?: SohoButtonComponent;
+  @ViewChild('tertiaryGenerative', { static: true }) tertiaryGenerative?: SohoButtonComponent;
+  @ViewChild('iconGenerative', { static: true }) iconGenerative?: SohoButtonComponent;
+
   constructor() { }
   ngOnInit() { }
   toggleHello() {
     this.shouldSayHi = !this.shouldSayHi;
+  }
+
+  performAnimation() {
+    this.primaryGenerative?.performAnimation(5000);
+    this.tertiaryGenerative?.performAnimation(5000);
+    this.iconGenerative?.performAnimation(5000);
+  }
+
+  startAnimation() {
+    this.primaryGenerative?.startAnimation();
+    this.tertiaryGenerative?.startAnimation();
+    this.iconGenerative?.startAnimation();
+  }
+
+  stopAnimation() {
+    this.primaryGenerative?.stopAnimation();
+    this.tertiaryGenerative?.stopAnimation();
+    this.iconGenerative?.stopAnimation();
   }
 }