+++ docs/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## 17.6.0 Features
 
-- `[Lookup]` Added `placeholder` setting for filter searchfield. ([#8416](https://github.com/infor-design/enterprise/issues/8416))
+- `[Button]` Added typings for generative button and updated example page. ([EP#8541](https://github.com/infor-design/enterprise/issues/8541))
+- `[Lookup]` Added `placeholder` setting for filter searchfield. ([EP#8416](https://github.com/infor-design/enterprise/issues/8416))
 
 ## 17.6.0 Fixes